Move MessageConsumer and MessageProducer to jsonrpc package
